hydration-bot/requirements.txt
# Hydration Reminder Bot dependencies
# Pinned versions for reproducibility.

python-telegram-bot[job-queue]==21.3
asyncpg==0.29.0
python-dotenv==1.0.1
